moved segment lane iterators to segment extensions created LaneUtil
